Move App component into component directory
